Change run!{} macro to accept patterns instead of identifiers.
Lambdas support pattern matching for their parameters. This change
enables the run!{} macro to accept patterns instead of identifiers.

The code is a bit convoluted, because of Follow-set Ambiguity
Restrictions. It is not possible to match a `pat` fragment in front of
"<=". Therefore the macro now matches against a `tt` and then tries to
pass the result to a nested macro, which only accepts `pat`. This allows
for relatively readable error messages.