Implement explicit clone in run!{} using [a,b] syntax.
This enables the user to explicitly clone values in do-notation. That is
needed because otherwise one would not be able to carry such values
along in do-notation for more than 2 statements.

The reason is the parameter of `bind()`, which has to be `Fn`. If that
parameter could be `FnOnce`, this issue would not exist, because a clean
chain of moves would be possible. However, limiting the parameter for
`bind()` would prevent implementing `Bind` for many structures that
should be Monads, for instance `Vec`.

With `bind()` taking `Fn`, the only option is to clone the
to-be-captured values, either implicitly if they are `Copy`, or
explicitly. This could of course be done before, but the syntax was
convoluted. This change simplifies the syntax to a list of
variables that one can add in front of the individual statements in
do-notation.